+++ src/main/java/edu/ithaca/icook/Recipe.java
@@ -1,0 +1,4 @@
+package edu.ithaca.goosewillis.iCook;
+
+public class Recipe {
+}
