+++ src/test/java/edu/ithaca/icook/RecipeTest.java
@@ -1,0 +1,7 @@
+package edu.ithaca.goosewillis.iCook;
+
+public class RecipeTest {
+
+
+
+}
